# openapi.yaml
openapi: 3.0.0
info:
  title: API LMS - Plateforme de gestion d'école
  description: "Examen Mi-Semestre L3 Informatique - Système complet de gestion scolaire."
  version: 1.0.0
  contact:
    email: admin@ecole.com
servers:
  - url: http://localhost:8000
    description: Serveur Local de Développement
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Course:
      type: object
      required: [title, instructor_id]
      properties:
        id: { type: integer, example: 1 }
        title: { type: string, example: "Laravel Avancé" }
        description: { type: string, example: "Maîtriser les APIs REST" }
        price: { type: number, example: 50000 }
        duration_hours: { type: integer, example: 20 }
        instructor_id: { type: integer, example: 1 }
...
paths:
...
  /api/v1/courses:
    get:
      tags: [Courses]
      summary: Liste des cours
      parameters:
        - name: instructor_id
          in: query
          schema: { type: integer }
        - name: sort
          in: query
          schema: { type: string, example: "price" }
      responses:
        200: { description: OK }
    post:
      security: [{ bearerAuth: [] }]
      tags: [Courses]
      summary: Créer un cours
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Course' }
      responses:
        201: { description: Créé }
  /api/v1/courses/{id}:
    get:
      tags: [Courses]
      summary: Détail cours + instructor
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        200: { description: OK }
    put:
      security: [{ bearerAuth: [] }]
      tags: [Courses]
      summary: Modifier un cours
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Course' }
      responses:
        200: { description: OK }
    delete:
      security: [{ bearerAuth: [] }]
      tags: [Courses]
      summary: Supprimer un cours
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        204: { description: OK }

  # --- AI (OPENAI) ---
...
  /api/v1/courses/{id}/description:
    get:
      tags: [AI]
      summary: Voir la description générée
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        200: { description: OK }

  # --- INSTRUCTORS ---
...
  /api/v1/enrollments:
    get:
      security: [{ bearerAuth: [] }]
      tags: [Enrollments]
      summary: Liste des inscriptions
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, active, completed] }
        - name: payment_status
          in: query
          schema: { type: string, enum: [unpaid, paid] }
      responses:
        200: { description: OK }
    post:
      security: [{ bearerAuth: [] }]
      tags: [Enrollments]
      summary: Inscrire un étudiant
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                student_id: { type: integer }
                course_id: { type: integer }
      responses:
        201: { description: OK }
...